move functions to cryptography package
