+++ cmd/crypto-vault-cli/crypto-cli.go
@@ -1,13 +1,7 @@
 package main
 
 import (
-	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/rand"
 	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -15,203 +9,9 @@ import (
 	"path/filepath"
 
 	"github.com/spf13/cobra"
+
+	cryptography "crypto_vault_service/pkg/cryptography"
 )
-
-// AES Functions
-// GenerateRandomAESKey generates a random AES key of the specified size
-func generateRandomAESKey(keySize int) ([]byte, error) {
-	key := make([]byte, keySize)
-	_, err := rand.Read(key)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate AES key: %v", err)
-	}
-	return key, nil
-}
-
-// Pad data to make it a multiple of AES block size
-func pkcs7Pad(data []byte, blockSize int) []byte {
-	padding := blockSize - len(data)%blockSize
-	padded := append(data, bytes.Repeat([]byte{byte(padding)}, padding)...)
-	return padded
-}
-
-// Unpad data after decryption
-func pkcs7Unpad(data []byte, blockSize int) ([]byte, error) {
-	length := len(data)
-	padding := int(data[length-1])
-
-	if padding > length || padding > blockSize {
-		return nil, fmt.Errorf("invalid padding size")
-	}
-	return data[:length-padding], nil
-}
-
-// Encrypts data using AES in CBC mode
-func encryptAES(plainText, key []byte) ([]byte, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, err
-	}
-
-	plainText = pkcs7Pad(plainText, aes.BlockSize)
-
-	ciphertext := make([]byte, aes.BlockSize+len(plainText))
-	iv := ciphertext[:aes.BlockSize]
-	if _, err := rand.Read(iv); err != nil {
-		return nil, err
-	}
-
-	mode := cipher.NewCBCEncrypter(block, iv)
-	mode.CryptBlocks(ciphertext[aes.BlockSize:], plainText)
-
-	return ciphertext, nil
-}
-
-// Decrypts data using AES in CBC mode
-func decryptAES(ciphertext, key []byte) ([]byte, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(ciphertext) < aes.BlockSize {
-		return nil, fmt.Errorf("ciphertext too short")
-	}
-
-	iv := ciphertext[:aes.BlockSize]
-	ciphertext = ciphertext[aes.BlockSize:]
-
-	mode := cipher.NewCBCDecrypter(block, iv)
-	mode.CryptBlocks(ciphertext, ciphertext)
-
-	return pkcs7Unpad(ciphertext, aes.BlockSize)
-}
-
-// RSA Functions
-func generateRSAKeys(bits int) (*rsa.PrivateKey, *rsa.PublicKey, error) {
-	privateKey, err := rsa.GenerateKey(rand.Reader, bits)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to generate RSA keys: %v", err)
-	}
-
-	publicKey := &privateKey.PublicKey
-	return privateKey, publicKey, nil
-}
-
-func savePrivateKeyToFile(privateKey *rsa.PrivateKey, filename string) error {
-	privKeyBytes := x509.MarshalPKCS1PrivateKey(privateKey)
-	privKeyPem := &pem.Block{
-		Type:  "RSA PRIVATE KEY",
-		Bytes: privKeyBytes,
-	}
-
-	file, err := os.Create(filename)
-	if err != nil {
-		return fmt.Errorf("failed to create private key file: %v", err)
-	}
-	defer file.Close()
-
-	err = pem.Encode(file, privKeyPem)
-	if err != nil {
-		return fmt.Errorf("failed to encode private key: %v", err)
-	}
-
-	return nil
-}
-
-func savePublicKeyToFile(publicKey *rsa.PublicKey, filename string) error {
-	pubKeyBytes, err := x509.MarshalPKIXPublicKey(publicKey)
-	if err != nil {
-		return fmt.Errorf("failed to marshal public key: %v", err)
-	}
-
-	pubKeyPem := &pem.Block{
-		Type:  "PUBLIC KEY",
-		Bytes: pubKeyBytes,
-	}
-
-	file, err := os.Create(filename)
-	if err != nil {
-		return fmt.Errorf("failed to create public key file: %v", err)
-	}
-	defer file.Close()
-
-	err = pem.Encode(file, pubKeyPem)
-	if err != nil {
-		return fmt.Errorf("failed to encode public key: %v", err)
-	}
-
-	return nil
-}
-
-// Encrypt data with RSA public key
-func encryptWithRSA(publicKey *rsa.PublicKey, data []byte) ([]byte, error) {
-	encryptedData, err := rsa.EncryptPKCS1v15(rand.Reader, publicKey, data)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encrypt data: %v", err)
-	}
-	return encryptedData, nil
-}
-
-// Decrypt data with RSA private key
-func decryptWithRSA(privateKey *rsa.PrivateKey, encryptedData []byte) ([]byte, error) {
-	decryptedData, err := rsa.DecryptPKCS1v15(rand.Reader, privateKey, encryptedData)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decrypt data: %v", err)
-	}
-	return decryptedData, nil
-}
-
-// File Operations
-func readFile(filePath string) ([]byte, error) {
-	return ioutil.ReadFile(filePath)
-}
-
-func writeFile(filePath string, data []byte) error {
-	return ioutil.WriteFile(filePath, data, 0644)
-}
-
-// AES Command
-
-// Read RSA private key from PEM file
-func readPrivateKey(privateKeyPath string) (*rsa.PrivateKey, error) {
-	privKeyPEM, err := ioutil.ReadFile(privateKeyPath)
-	if err != nil {
-		return nil, fmt.Errorf("unable to read private key file: %v", err)
-	}
-
-	block, _ := pem.Decode(privKeyPEM)
-	if block == nil {
-		return nil, fmt.Errorf("failed to parse PEM block containing the private key")
-	}
-
-	privateKey, err := x509.ParsePKCS1PrivateKey(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("unable to parse private key: %v", err)
-	}
-
-	return privateKey, nil
-}
-
-// Read RSA public key from PEM file
-func readPublicKey(publicKeyPath string) (*rsa.PublicKey, error) {
-	pubKeyPEM, err := ioutil.ReadFile(publicKeyPath)
-	if err != nil {
-		return nil, fmt.Errorf("unable to read public key file: %v", err)
-	}
-
-	block, _ := pem.Decode(pubKeyPEM)
-	if block == nil {
-		return nil, fmt.Errorf("failed to parse PEM block containing the public key")
-	}
-
-	publicKey, err := x509.ParsePKCS1PublicKey(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("unable to parse public key: %v", err)
-	}
-
-	return publicKey, nil
-}
 
 // Encrypts a file using AES and saves the encryption key
 func encryptAESCmd(cmd *cobra.Command, args []string) {
@@ -226,24 +26,24 @@ func encryptAESCmd(cmd *cobra.Command, args []string) {
 	}
 
 	// Generate AES Key
-	key, err := generateRandomAESKey(keySize)
+	key, err := cryptography.GenerateRandomAESKey(keySize)
 	if err != nil {
 		log.Fatalf("Error generating AES key: %v\n", err)
 	}
 
 	// Encrypt the file
-	plainText, err := readFile(inputFile)
+	plainText, err := cryptography.ReadFile(inputFile)
 	if err != nil {
 		log.Fatalf("Error reading input file: %v\n", err)
 	}
 
-	encryptedData, err := encryptAES(plainText, key)
+	encryptedData, err := cryptography.EncryptAES(plainText, key)
 	if err != nil {
 		log.Fatalf("Error encrypting data: %v\n", err)
 	}
 
 	// Save encrypted file
-	err = writeFile(outputFile, encryptedData)
+	err = cryptography.WriteFile(outputFile, encryptedData)
 	if err != nil {
 		log.Fatalf("Error writing encrypted file: %v\n", err)
 	}
@@ -251,7 +51,7 @@ func encryptAESCmd(cmd *cobra.Command, args []string) {
 
 	// Save the AES key to the specified key directory
 	keyFilePath := filepath.Join(keyDir, "encryption_key.bin")
-	err = writeFile(keyFilePath, key)
+	err = cryptography.WriteFile(keyFilePath, key)
 	if err != nil {
 		log.Fatalf("Error writing AES key to file: %v\n", err)
 	}
@@ -276,18 +76,18 @@ func decryptAESCmd(cmd *cobra.Command, args []string) {
 	}
 
 	// Decrypt the file
-	encryptedData, err := readFile(inputFile)
+	encryptedData, err := cryptography.ReadFile(inputFile)
 	if err != nil {
 		log.Fatalf("Error reading encrypted file: %v\n", err)
 	}
 
-	decryptedData, err := decryptAES(encryptedData, key)
+	decryptedData, err := cryptography.DecryptAES(encryptedData, key)
 	if err != nil {
 		log.Fatalf("Error decrypting data: %v\n", err)
 	}
 
 	// Save decrypted file
-	err = writeFile(outputFile, decryptedData)
+	err = cryptography.WriteFile(outputFile, decryptedData)
 	if err != nil {
 		log.Fatalf("Error writing decrypted file: %v\n", err)
 	}
@@ -305,43 +105,43 @@ func encryptRSACmd(cmd *cobra.Command, args []string) {
 	var err error
 	if publicKeyPath == "" {
 		// Generate RSA keys
-		privateKey, pubKey, genErr := generateRSAKeys(2048)
+		privateKey, pubKey, genErr := cryptography.GenerateRSAKeys(2048)
 		if genErr != nil {
 			log.Fatalf("Error generating RSA keys: %v\n", genErr)
 		}
 		publicKey = pubKey
 
 		// Optionally save the private and public keys
-		err = savePrivateKeyToFile(privateKey, "data/private_key.pem")
+		err = cryptography.SavePrivateKeyToFile(privateKey, "data/private_key.pem")
 		if err != nil {
 			log.Fatalf("Error saving private key: %v\n", err)
 		}
-		err = savePublicKeyToFile(publicKey, "data/public_key.pem")
+		err = cryptography.SavePublicKeyToFile(publicKey, "data/public_key.pem")
 		if err != nil {
 			log.Fatalf("Error saving public key: %v\n", err)
 		}
 		fmt.Println("Generated and saved RSA keys.")
 	} else {
 		// Read the provided public key
-		publicKey, err = readPublicKey(publicKeyPath)
+		publicKey, err = cryptography.ReadPublicKey(publicKeyPath)
 		if err != nil {
 			log.Fatalf("Error reading public key: %v\n", err)
 		}
 	}
 
 	// Encrypt the file
-	plainText, err := readFile(inputFile)
+	plainText, err := cryptography.ReadFile(inputFile)
 	if err != nil {
 		log.Fatalf("Error reading input file: %v\n", err)
 	}
 
-	encryptedData, err := encryptWithRSA(publicKey, plainText)
+	encryptedData, err := cryptography.EncryptWithRSA(publicKey, plainText)
 	if err != nil {
 		log.Fatalf("Error encrypting data: %v\n", err)
 	}
 
 	// Save encrypted file
-	err = writeFile(outputFile, encryptedData)
+	err = cryptography.WriteFile(outputFile, encryptedData)
 	if err != nil {
 		log.Fatalf("Error writing encrypted file: %v\n", err)
 	}
@@ -358,39 +158,39 @@ func decryptRSACmd(cmd *cobra.Command, args []string) {
 	var err error
 	if privateKeyPath == "" {
 		// Generate RSA keys
-		privKey, _, genErr := generateRSAKeys(2048)
+		privKey, _, genErr := cryptography.GenerateRSAKeys(2048)
 		if genErr != nil {
 			log.Fatalf("Error generating RSA keys: %v\n", genErr)
 		}
 		privateKey = privKey
 
 		// Optionally save the private and public keys
-		err = savePrivateKeyToFile(privateKey, "private_key.pem")
+		err = cryptography.SavePrivateKeyToFile(privateKey, "private_key.pem")
 		if err != nil {
 			log.Fatalf("Error saving private key: %v\n", err)
 		}
 		fmt.Println("Generated and saved private key.")
 	} else {
 		// Read the provided private key
-		privateKey, err = readPrivateKey(privateKeyPath)
+		privateKey, err = cryptography.ReadPrivateKey(privateKeyPath)
 		if err != nil {
 			log.Fatalf("Error reading private key: %v\n", err)
 		}
 	}
 
 	// Decrypt the file
-	encryptedData, err := readFile(inputFile)
+	encryptedData, err := cryptography.ReadFile(inputFile)
 	if err != nil {
 		log.Fatalf("Error reading encrypted file: %v\n", err)
 	}
 
-	decryptedData, err := decryptWithRSA(privateKey, encryptedData)
+	decryptedData, err := cryptography.DecryptWithRSA(privateKey, encryptedData)
 	if err != nil {
 		log.Fatalf("Error decrypting data: %v\n", err)
 	}
 
 	// Save decrypted file
-	err = writeFile(outputFile, decryptedData)
+	err = cryptography.WriteFile(outputFile, decryptedData)
 	if err != nil {
 		log.Fatalf("Error writing decrypted file: %v\n", err)
 	}
