+++ pkg/cryptography/hashing.go
@@ -1,0 +1,1 @@
+package cryptography
